Update deprecated methods from PlaceHolderAPI
