+++ pom.xml
@@ -146,7 +146,7 @@
         <dependency>
             <groupId>me.clip</groupId>
             <artifactId>placeholderapi</artifactId>
-            <version>2.10.2</version>
+            <version>2.10.6</version>
         </dependency>
     </dependencies>
 </project>
